Fix potential LUI scripting crash

+++ src/client/component/server_list.cpp
@@ -111,8 +111,8 @@ namespace server_list
 			return static_cast<int>(servers.size() - server_list_index);
 		}
 
-		const char* ui_feeder_item_text(int /*localClientNum*/, void* /*a2*/, void* /*a3*/, const size_t index,
-		                                const size_t column)
+		const char* ui_feeder_item_text(int /*localClientNum*/, void* /*a2*/, void* /*a3*/, const int index,
+		                                const int column)
 		{
 			std::lock_guard<std::mutex> _(mutex);
